Safari-611.1.10/LayoutTests/fast/forms/ios/autocapitalize-words-expected.txt

This test verifies that typing in editable areas with autocapitalize='words' causes every word to be capitalized. To test manually, type into both elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.value is "Do Re"
PASS contenteditable.innerText is "Mi Fa\nSo"
PASS successfullyParsed is true

TEST COMPLETE
